requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
psycopg2-binary
numpy
pydantic
atproto
sentence-transformers
google-generativeai
pgvector
itsdangerous
python-multipart
pandas